+++ src/main/kotlin/com/github/alanxtl/wslpathconverter/MyBundle.kt
@@ -1,0 +1,13 @@
+package com.github.alanxtl.wslpathconverter
+
+import com.intellij.DynamicBundle
+import org.jetbrains.annotations.PropertyKey
+
+private const val BUNDLE = "messages.MyBundle"
+
+object MyBundle : DynamicBundle(BUNDLE) {
+
+    fun message(@PropertyKey(resourceBundle = BUNDLE) key: String, vararg params: Any): String {
+        return getMessage(key, *params)
+    }
+}